% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hyreg}
\alias{hyreg}
\title{Perform hybrid tobit-logit regression}
\usage{
hyreg(formula.tobit, formula.discrete, data.tobit, data.discrete,
  start.beta = NULL, start.sigma = NULL, start.theta = 1,
  left = -1, method = "Nelder-Mead")
}
\arguments{
\item{formula.tobit}{a regression formula describing the relationship between the tobit response and the covariates}

\item{formula.discrete}{a regression formula describing the relationship between the bernoulli response and the covariates}

\item{data.tobit}{the data.frame containing the tobit responses and covariates}

\item{data.discrete}{the data.frame containing the bernoulli responses and covariates}

\item{start.beta}{a numeric vector of starting values for beta. If not specified, start.beta is taken from a non-hybrid tobit model}

\item{start.sigma}{a numeric starting value for sigma. If not specified, start.sigma is also taken from a non-hybrid tobit model}

\item{start.theta}{starting value for theta.}

\item{left}{a number specifying where left-censoring occurred}

\item{method}{a string specifying the optimization routine to be used by optim}
}
\value{
a list containing the following parameter estimates (from maximum likelihood):\cr
\item{beta}{the regression coefficients}
\item{sigma}{the standard deviation of the censored normal distribution}
\item{theta}{the multiplicative factor relating the two sets of regression coefficients}
}
\description{
This function gets estimates of beta, sigma, and theta in the hybrid tobit-logit
model. That is, when the logit betas are a scalar multiple of the tobit betas.
}
